Tic Tac Toe App
In this application, we created a TicTacToeApp An android game that is designed to play between two players. Opens with animation by android. The game requires two players to sign up. Will require both the user to submit their name. Both the players get a chance one by one. The first chance is Player 1. If there is a winner after 9 chances, then the textView will display the Winner's Name. If there is a draw, then the textView will display that the Match has been a tie.